ota: tell callback whether it's the last block
The "get block" callback functions now take an additional
param, indicating if it's the last block of the GET.

This is needed to know when all blocks have been downloaded
during OTA update. Previously, we relied on the manifest size
to know how many blocks to download, but there is currently
a bug in the Golioth backend that results in the manifest
size not being reported accurately (does not match binary size),
so we have to use is_last flag to know when all blocks
have been downloaded.

Signed-off-by: Nick Miller <nick@golioth.io>