test: add settings test
Relies heavily on Golioth API to create/delete/read settings.
Test firmware modified to receive the TEST_SETTING setting
and set the value received in LightDB state so it can be
read via API and confirmed to match the original value.

Leaving the debug prints in verify.py since this is a new test
and I'm not sure how stable/flaky it will be in CI.

Signed-off-by: Nick Miller <nick@golioth.io>