test: after connect, wait for empty req queue
Now that we have a settings callback, there is an additional
POST to .c/status that happens shortly after connecting
to Golioth, which is throwing off the timing of tests
that run just after connecting, causing test assertion failures.

To resolve the timing sensitive tests, we will wait for the
request queue to be completely empty before proceeding, which
will give enough time for the post to .c/status to complete.

Signed-off-by: Nick Miller <nick@golioth.io>