settings: no error in case callback already set
Previously, we would return an error if the user tries
to set the settings callback more than once. This was
causing an error in the test firmware, which creates
more than one client.

Returning an error is not really necessary, and instead
the callback can simply be overwritten.

Ideally, we need to store the callback inside the
client struct instead of as a static global, which would
have prevented this problem. Added a TODO to settings and
rpc to address this in the future.

Signed-off-by: Nick Miller <nick@golioth.io>