settings: move global data to client struct
Previously, golioth_settings relied on global data for storing
the user's callback. However, this does not work very well
in case of multiple clients, or in cases where the client
is destroyed and recreated (such as in test code).

By storing the data in the client struct, there can be settings
data stored per client in a thread-safe manner.

Signed-off-by: Nick Miller <nick@golioth.io>