settings: simpler user API, redesign
Redesigned the front end API of golioth_settings.h (this is
a breaking change).

The goal was to make the user's settings callback much
smaller, as the prior version resulted in a lot of user
boilerplate code to string compare the key, verify the
type is correct, verify the range is correct, etc.

The new API is oriented around registering callbacks for
individual settings of a known type.

Most of the basic checks are now happening inside of
the SDK (key exists/known, value type, range check),
resulting in much simpler user callback functions.

Signed-off-by: Nick Miller <nick@golioth.io>