MTB: README: requires MTB 2.4.0
The newest version of MTB (3.0.0) is not compatible
with the example code (mostly Makefile and build system
differences).

The differences between MTB 2.4.0 and 3.0.0 are entirely in
the build system and don't affect the example C code.

For now, we will keep this example pinned to 2.4.0 for simplicity
and ease of development. We link to the "official" MTB Golioth
example which supports both 2.4.0 and 3.0.0.

Signed-off-by: Nick Miller <nick@golioth.io>