coap_client: detect handshake failures
Previously, there was no way to differentiate between a handshake
failure and a request timeout. The "Maybe your PSK-ID or PSK is
incorrect?" message would display regardless, which is confusing
in the timeout case, since normally the credentials are correct.

Libcoap has an internal state machine that will be in the "handshake"
state, so use that to detect when the handshake is failed.

Signed-off-by: Nick Miller <nick@golioth.io>