libcoap: update to tip of develop
This pulls in a recent libcoap feature with support for
the COAP_EVENT_MSG_RETRANSMITTED event, which will be
used for tracking link quality and server responsiveness
issues.

Reordered the includes in the esp-idf and MTB ports to get it build.
Matched the order of includes from libcoap's coap3/coap.h.in.

Signed-off-by: Nick Miller <nick@golioth.io>