golioth_basics: GLTH_LOGX instead of golioth_log
The golioth_log* functions, which send logs to Golioth, are
not very useful in their current form since they are not
tied into the log output of the platform (i.e. a user must
explicitly send a log to Golioth, instead of it happening
automatically when printf is called).

The long-term solution is to have all logs/device output
sent to Golioth if enabled, so GLTH_LOGX statements are
preferred.

Eventually the golioth_log* functions will be
deprecated, so this is the first step towards that.

Signed-off-by: Nick Miller <nick@golioth.io>