README: add feature x platform table
The new remote shell feature does not yet work on ModusToolbox,
so it's a good excuse to make a table in the README with a list
of features of the core SDK and whether they're supported on
each platform.

Also, a capitalization fix and an update to the last tested
version and date of ModusToolbox.

Signed-off-by: Nick Miller <nick@golioth.io>