replace FreeRTOS tasks with golioth_sys_thread_t
There were three tasks to replace: coap_client, fw_update, and
remote_shell.

Also, removed the "task stack high watermark" function since
this is a FreeRTOS-specific function. It's not important to
the function of the SDK, and is only used in the esp-idf "test"
project, so in that project we just call FreeRTOS directly to
get the stack HWM.

Signed-off-by: Nick Miller <nick@golioth.io>