coap_client: purge request queue on destroy
For POST requests, we make a dynamically-allocated
copy of the user's payload which gets
freed when the request is popped off the queue.

If the client is destroyed while requests are still in
the queue, then it's possible that memory will leak,
so iterate through the entire queue and free the memory
of any POST requests in that case.

Signed-off-by: Nick Miller <nick@golioth.io>