config: user config file path override
Previously, we assumed if the user wanted to override
config, there had to be a file named "golioth_user_config.h" that
was on the include path.

This was not easy to use, because user config files are usually
outside the include path of the golioth_sdk library. By allowing
the user to define the path to the user config file, it can
be set from the outside.

Signed-off-by: Nick Miller <nick@golioth.io>