ci: run test_esp32s3.yml every night
Currently, the workflow only runs on PRs and merges to main,
meaning if there is no activity in this repo, it's possible
a change to Golioth servers could have broken the end-to-end
test.

Running a nightly test on a fixed schedule should help catch
any serious issues within 24 hours, regardless of whether
a PR has been created.

Signed-off-by: Nick Miller <nick@golioth.io>