docs: add Style_Guide.md
The intent is to be explicit about the style guidelines that
matter to this project without be overly constraining.

The fallback style is Linux kernel style. We deviate from Linux
kernel style in several significant ways, but these are mostly
for conformance with existing code, not because our style is any
better.

Signed-off-by: Nick Miller <nick@golioth.io>