golioth_debug: GLTH_LOGX macro now logs to Golioth
Added capability for the default GLTH_LOGX macro to log to
both stdout and to Golioth. It uses the new function
`golioth_debug_printf` to serialize the string to a dynamically
allocated buffer and asynchronously log to Golioth.

This is more convenient than the golioth_log_X functions, which
require the user to allocate a buffer and sprintf to it before
logging to Golioth. Now they can simply use GLTH_LOGX in a
printf style, without needing a buffer or sprintf.

Special care is taken to prevent infinite log loops, in case an
error is logged while trying to log to Golioth.

Signed-off-by: Nick Miller <nick@golioth.io>