port,mtb: override GLTH_LOGX to avoid 64-bit num
The basic ModusToolbox libc does not have good support for
use 64-bit formatters (like PRIu64) in the printf family
of functions. MTB does support it with newlib, but that
is a larger library that most MTB examples/applications
don't use, so we should avoid it too.

For the MTB port, we override GLTH_LOGX to use 32-bit timestamps.
This is necessary, otherwise the log output become 30% garbled
mess.

Signed-off-by: Nick Miller <nick@golioth.io>