examples,mtb: define creds in credentials.inc
Previously, the WiFi and Golioth credentials were hard-coded in
golioth_main.h. This was problematic for source control, because
it has to be modified with valid credentials to test, and it easy
to forget to revert those changes, or to accidentally commit
them to source control.

Instead, we ask the user to define the credentials in a file
named credentials.inc, which gets included in golioth_main.h
and .gitignore'd.

Signed-off-by: Nick Miller <nick@golioth.io>