config: add option to auto log to Golioth
The option CONFIG_GOLIOTH_AUTO_LOG_TO_CLOUD was added to
configure whether GLTH_LOGX statements should automatically
upload the log to Golioth. It is disabled by default
due to the high amount of transmitted data when enabled.

This configuration item will be the default used for enabling
or disabling the automatic cloud log feature, but it can
also be overridden at runtime (which might be useful for
runtime enabling of logging, for troubleshooting purposes).

Signed-off-by: Nick Miller <nick@golioth.io>