esp-idf: move libcoap to separate component
Before this change, libcoap was built as part of the
libgolioth_sdk.a component. When using `idf.py size-components`,
it was hard to tell how much was contributed from the Golioth
SDK, and how much was from libcoap.

Splitting into separate components allows us to analyze
the static DRAM/flash usage of the SDK and libcoap
components separately, which will be useful for future optimization.

Signed-off-by: Nick Miller <nick@golioth.io>