port/esp_idf: use OTA_SIZE_UNKNOWN in esp_ota_begin
Before this change, we were passing the total image size from
the manifest to esp_ota_begin. This erases the minimum number
of flash blocks needed to store the image.

Looking forward to image compression and diffs, we won't know
from the manifest how large the final uncompress/dediff'd file
will be, so use OTA_SIZE_UNKNOWN, which will cause the entire
flash partition to be erased.

This does not have a noticeable impact of flash erase times
compared to using the image size directly.

Signed-off-by: Nick Miller <nick@golioth.io>