golioth_fw_update: support for block decompression
Components in the manifest have a new `is_compressed` field
to indicate whether the component is compressed with the
heatshrink algorithm.

The firmware update download loop will decompress blocks as
they are received using the heatshrink library if `is_compressed`
is set to `true`.

The window size and lookahead size for decompression are set to
8 and 4 respectively, which seemed like a reasonable default
for a resource-contrained device.
In practice, it achieves ~20% compression ratio.

Since the OTA block compression feature is not yet supported
on the Golioth cloud, the `is_compressed` field is hard-coded
to `false`, effectively disabling decompression for now.

Once supported by the cloud, we will need to revisit the code
in golioth_ota.c to set `is_compressed` based on the manifest.

Signed-off-by: Nick Miller <nick@golioth.io>