port/linux: keep sleeping if EINTR
Before this change, golioth_sys_msleep was not sleeping
long enough because there was a signal interrupting it
from the 9s keepalive timer.

Now, we use nanosleep, which will still be interrupted, but
will tell us how much longer we need to sleep, so we sleep
in a loop until we're done sleeping or some other non-EINTR
signal happens.

Signed-off-by: Nick Miller <nick@golioth.io>