golioth_coap_client: use Accept for GET/OBSERVE
Before this change, we used a Content-Format option
on GET requests to specify the preferred format of the
response. But this is not correct. The spec states
(https://www.rfc-editor.org/rfc/rfc7252#section-5.10.3):

```
The Content-Format Option indicates the representation
format of the message payload.
```

The server did not use the Content-Format, and replied with
text/plain by default (which is actually in JSON format).
Our SDK only parses JSON, so we did not detect this problem.
However, it would have been an issue if another format (e.g. CBOR)
were needed.

We now use the Accept option to specify the desired response
format, per the spec.

Signed-off-by: Nick Miller <nick@golioth.io>