.github: use --release-tags when deleting a release
Before this change, the goliothctl `release delete` command
did not include the `--release-tags` option. However, the release
was still getting deleted because it was the only release in
that project (i.e. it worked by accident).

This change explicitly sets `--release-tags`, which is required
if there is ever more than one release in the project. It also
nicely mirrors the `release create` command.

Signed-off-by: Nick Miller <nick@golioth.io>