coap_client: refactor client connect event
Before this change, the code handling connection events was odd to read,
because the logic made it appear as "if not connected, then treat as
a connection event" which at first glance seems incorrect.
The logic was correct, since we are handling the transition from not
connected to connected, but that was not obvious in the code. Also, there
was an extraneous check for !client->session_connected.

The refactor makes the intention more clear and reduces duplicated code.

Signed-off-by: Nick Miller <nick@golioth.io>