.github: run 4.4.2 build in a separate job
When trying to run the 4.4.2 build after a 5.0 build, there was an
error with the version of CMake found on PATH:

/bin/sh: 1: /opt/esp/tools/cmake/3.24.0/bin/cmake: not found

This is because the IDF 5.0 tools are first in PATH, and get used
in the 4.4.2 build, which does not work.

Running the 4.4.2 build in a separate job works around this.

Signed-off-by: Nick Miller <nick@golioth.io>