docs: Add ModusToolbox to integration guide
Shows the general procedure for adding Golioth to an existing MTB project.

It's not a full integration guide, but more of a "fuzzy" integration guide,
showing general steps, but leaving out some of the details of how to add
the MCUboot project and miscellaneous python and Makefile scripts for
configuration of the flash map. This was done to keep the integration guide
short and readable, and also because it's recommended for users to start
from the Golioth example instead of trying to add Golioth SDK to an
existing project.

Signed-off-by: Nick Miller <nick@golioth.io>