esp_idf,shell: verify key before accessing NVS
Before this change, if an unknown key was entered in the settings command,
the string "unknown" would be used as the key when reading/writing NVS
on the device. There was no error returned to the user
(appeared successful). This allowed users to enter invalid keys
(like "wifi/password", a common mistake), and be led to believe the key
was set successfully, when in fact it was not.

This changes moves the known keys into a lookup table, then returns an
error if the key entered by the user isn't in the lookup table.

Signed-off-by: Nick Miller <nick@golioth.io>