coap_client: include time.h header
This header is needed for `time_t time(time_t *tloc);` prototype. Fix build
time compilation on platforms that do not pull `time.h` implicitly.

Note that `srand(time(&t));` is only secure when there is correct wall
clock time already provided in the system. This is true in most cases for
general purpose operating systems, like Linux, where SDK is used as part of
user process started at the end of boot procedure. In case of MCU OSes,
like FreeRTOS and Zephyr, it is much more likely that time is not correctly
synchronized. As a result `time()` possibly returns time from boot. Such
time is relatively deterministic and this can result in selecting similar
or the same seed for PRNG, which seems to be security issue.

So in summary, fix build time issue for now, but keep in mind security
issues related to the use of `srand(time(&t));` invocation and relying on
PRNG from that point.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>