fw_update: move decompression logic to separate file
Created new files golioth_decompressor.h/.c, to clean up the structure
of download_and_write_flash().

The main idea is that OTA block data is processed via a chain of
function calls. With this commit, the chain is:

block data -> decompress -> handle_block

This restructuring will make it easier to add a patch step in the
processing chain in later commits.

Signed-off-by: Nick Miller <nick@golioth.io>