fw_update: use static heatshrink decoder
The heatshrink submodule commit was updated, which pulls in configuration
changes to disable dynamic allocation and set the static input buffer
size to 1024. This was done because dynamic allocation of the decoder
makes the SDK fw_update code more complicated and prone to leaks, and
because the appropriate input buffer size is known at compile time.

As a result of this configuration change, the decompression code in
golioth_fw_update.c needed to be updated to remove the alloc/free calls
and to statically allocate a heatshrink_decoder.

Removed the runtime "is_compressed" flag from golioth_ota_component_t.
Instead, we are now relying on compile-time configuration
CONFIG_GOLIOTH_OTA_DECOMPRESSION_ENABLE to indicate whether decompression
should be done. This was done because we want to avoid statically
allocating a heatshrink_decoder in the case where decompression is
disabled, and this needs to be known at compile time, so the
is_compressed flag is no longer useful.

The #if/#else guards make the golioth_fw_update.c code slightly more
complicated to read, but it is only temporary. There will be a refactor
in the next few commits which moves the decompression logic to a
separate file, and #if/#else guards will be removed.

Signed-off-by: Nick Miller <nick@golioth.io>