fw_update: connect output of download to decompress
Instead of calling decompress_and_handle() directly in
download_and_write_flash(), we will instead have download_block() call
decompress_and_handle() when a block has been downloaded.

This allows the "output" of download_block() to be connected to the
"input" of decompress_and_handle(), via usage of a function pointer.

By using this method of connecting output to input, it is possible
to create a chain of data processing stages, where the output of one
stage is connected to the input of the next.

This kind of processing is useful for fw_update because it:

* Decouples processing stages
* Allows the top-level download_and_write_flash() to be agnostic of the
  processing stages, and instead simply invoke the first function in the
  processing chain (download_block()).
* Allows other processing stages to be inserted into the chain.

The last bullet is the main motivator for this change, because in future
commits, it is planned to:

* Split the decompress_and_handle() stage into two separate stages:
  decompress() and handle()
* Add a new processing stage for "patch", to implement the OTA delta
  feature.

Signed-off-by: Nick Miller <nick@golioth.io>