fw_update: split decompress_and_handle() into separate funcs
Using the "chaining" method introduced in the previous commit (i.e.
connecting the output of one processing stage to the input of the next via
function pointer), this commit splits the decompress_and_handle() stage
into two separate stages: decompress() and handle_block(), where the
output of decompress() is connected to the input to handle_block() via
function pointer.

Handling of a block is now a separate processing stage, and thus has
a separate context struct and functions, similar to the download and
decompress stages.

Signed-off-by: Nick Miller <nick@golioth.io>