fw_update: new container struct block_process_ctx_t
This commit introduces a new struct block_process_ctx_t, which is
a container for the individual elements of the block processing chain,
namely download, decompress, and handle_block.

The wiring of the processing chain is encapsulated in
block_processor_init().

This container struct simplifies the code in download_and_write_flash(),
which no longer needs to instantiate the individual structs, initialize
them, or wire them together.

Signed-off-by: Nick Miller <nick@golioth.io>