fw_update: move block processing code to separate .h/.c files
Before this change, there was a lot of code in golioth_fw_update related
to processing OTA blocks.

This code has been cut-and-pasted to golioth_fw_block_process.h/.c to
improve readability of both the golioth_fw_update code and the block
processing code.

These new files are for internal use only, not intended to be exposed
publicly to the SDK user.

Signed-off-by: Nick Miller <nick@golioth.io>