heatshrink: update to latest develop branch
Updated the heatshrink submodule to the tip of the develop branch.

The primary motivation for doing this is that configuration in
heatshrink_config.h can now be overridden. This is planned
to be used to disable dyanmic allocation in heatshrink (setting
HEATSHRINK_DYNAMIC_ALLOC to 0).

The include and source file paths changed a bit, so CMakeLists.txt files
have been updated accordingly.

Signed-off-by: Nick Miller <nick@golioth.io>