fw_block_processor: use static heatshrink decoder
Instead of dynamically allocating the heatshrink decoder, we will
statically allocate it instead. Using a static decoder make the fw
update code simpler and avoids potential memory leaks.

This requires heatshrink to be compiled with HEATSHRINK_DYNAMIC_ALLOC=0,
so CMakeLists.txt/Makefiles have been modified to set this.

Prior to this commit (when dynamic alloc decoder was used), the input
buffer was set to 1024 bytes. There was no reason for it to be this large.
And now that we are using a static decoder, allocating 1024 bytes
statically could be a problem if the decoder is on the stack.
Instead, the default input buffer size of 32 bytes will be used, which
makes this safer for stack allocations and generally requires less
RAM. As a result, the logic in decompress() has been modified to
make multiple calls into heatshrink_decoder_sink() for each block.

Signed-off-by: Nick Miller <nick@golioth.io>