coap_client: allow for no block2 option in block response
If a CoAP client makes a block2 request for a resource that is less
than the size of a block (e.g. < 1024 bytes), the Golioth server
will not include the BLOCK2 option in the response.

Logic for handling block2 requests has been updated accordingly, to
allow for a possible NULL return from
coap_check_option(COAP_OPTION_BLOCK2).

Signed-off-by: Nick Miller <nick@golioth.io>