port/esp-idf: remove fw_update_header_valid()
With the streaming-based approach of the block processor, sometimes the
full IDF header is not available in the first chunk.

This check is not strictly necessary, since there is a validation step
at the end (esp_ota_end()) which would catch any format/checksum errors.

Fixing the code so we wait long enough for the full header to be received
before calling fw_update_header_valid() would complicate things, and
given that the check isn't necessary, it's preferred to simply delete it.

Signed-off-by: Nick Miller <nick@golioth.io>