.github: deploy new OTA release from self-hosted runner
Previously, we relied on the repo secret GOLIOTH_API_KEY to be able to
deploy the OTA release from GitHub Actions in the cloud.

This is problematic in the case where there are multiple self-hosted
runners, each which might be using a different Golioth project with
different API keys.

To resolve, we will deploy the OTA release from the self-hosted runner,
where the runner-specific API key is already known and present in the file
$HOME/credentials_esp32s3.yml.

This will allow for deleting the repo secret GOLIOTH_API_KEY.

Signed-off-by: Nick Miller <nick@golioth.io>