esp_idf/test: increase connect timeout to 2x resp timeout
test_connects_to_golioth() often fails due to a known bug on
the backend that can cause the DTLS handshake to timeout (GB-564).

In a prior commit, the macro TEST_RESPONSE_TIMEOUT_S was introduced
which sets the connect timeout to 2x + 1 seconds, where x is the
timeout for a single request. This allows for one timeout to occur
without failing the test.

It was intended that the test_connects_to_golioth() test use this macro,
but it was overlooked/missed in the prior commit.

Signed-off-by: Nick Miller <nick@golioth.io>