README: clarify the support for OTA on Linux
Before this change, the features table indicated that Linux
"OTA FW Update" was coming :soon:. However, other OTA features like
compression and delta updates were already marked as supported, which
was confusing, since one would expect the baseline FW update
functionality to be a prerequisite for the other features.

This marks "OTA FW Update" as supported for Linux, but with a note that
explains it is only partially implemented.

Fixes #101

Signed-off-by: Nick Miller <nick.miller.83@gmail.com>