scripts: Use git-clang-format in check_clang_format.sh
Previously, if a file had been modified then check_clang_format.sh
would check the entire file for code formatting, including the unchanged
portions of the file. git-clang-format is more precise, and will only
check those lines which are in the diff. This also allows us to use
existing tooling instead of custom.

Signed-off-by: Sam Friedman <sam@golioth.io>