zephyr: common: add GOLIOTH_SAMPLE_FW_VERSION
Use this Kconfig symbol to set the firmware version number. It will be used
by the SDK as the string version number for firmware upgrades and by
MCUboot for the slot version number.

Zephyr and NCS use different symbols to pass the version number to MCUboot.
This commit automatically assigns the correct symbol using
GOLIOTH_SAMPLE_FW_VERSION.

Signed-off-by: Mike Szczys <mike@golioth.io>