samples: add connection test
Adds a simple connection test which determines if the DUT can connect
to the Golioth backend. This is useful as a smoke test.

As part of this commit, we also add the beginnings of a HIL test
framework. The framework is intended to allow HIL tests to be
portable across different boards and platforms. Currently, only
the nRF52840DK running Zephyr is supported, and so not all of
the abstractions are in place. These abstractions will be added
in future commits as needed as we add support for additional
boards and platforms.

Signed-off-by: Sam Friedman <sam@golioth.io>