mbox: do not place functional code in assert()
The mbox module had some of its logic wrapped in calls to assert(). When
assertions are turned off, that logic was compiled out, and the mbox module
stopped working. assert() should always only be used for testing values,
and should never be used to wrap function calls, assignments, or other
statements with side effects.

Signed-off-by: Sam Friedman <sam@golioth.io>