examples: zephyr: s/hardcoded_credentials/sample_credentials
We will use the same API to get runtime PSK settings and hardcoded PSK
settings in the samples. This renames the APIs to reflect this. Right now,
hardcoded_credentials.c is the only implementation of the new API
(`golioth_sample_credentials_get()`), but we will add an implemntation
for runtime PSK settings in future commits.

Signed-off-by: Sam Friedman <sam@golioth.io>