examples: zephyr: common: add runtime psk config overlay
Adds a .conf file for enabling runtime settings for PSK and WiFi. This
can be passed as an argument to `west build` to enable the settings
with one simple command instead of adding the individual Kconfig options
to `prj.conf` (which still works). To use:

`west build -p -b <board> <app> --
    -DEXTRA_CONF_FILE=../common/runtime_psk.conf`

Signed-off-by: Sam Friedman <sam@golioth.io>