examples: zephyr: reduce heap alloc for esp32_devkitc_wrover
The ESP32-DevKitC-WROVER board overflowed the dram_0 segment when
credential auth was enabled. Reducing the RAM allocated to the system
heap alleviates the oveflow and doesn't seem to affect any other systems.
ESP32 has the ability to use dram_1 as an additional source of memory for
the system heap using the Kconfig symbol ESP_HEAP_MEM_POOL_REGION_1_SIZE
and this can be increased if more heap is needed.

Signed-off-by: Sam Friedman <sam@golioth.io>