ci: parallelize twister builds
Creates a reusable workflow for running twister build tests on
a single board, with either Zephyr or NCS. Uses this workflow
to run build tests on each of our CVBs. Because these are run
as separate steps in the caller workflow, they run in parallel
on separate runners, and complete faster than if they were
run sequentially.

Signed-off-by: Sam Friedman <sam@golioth.io>