ci: use minimal docker image for build test
The official Zephyr CI Docker image is rather large, and includes
many toolchains, modules, and other things that we don't need for
our own builds. Using our own, smaller Docker image in CI can cut
the time it takes to download and initialize the container from
about 2-2.5 minutes down to about 30 seconds. This container does
not come pre-installed with Zephyr's required python dependencies
so this also adds a step in the CI workflow to install them, which
is more robust anyway.

Signed-off-by: Sam Friedman <sam@golioth.io>