hil: add pytest plugin for boards
Adds a pytest plugin (`hil`) that provides a pytest fixture for
interacting with boards connected over serial. This provides an
abstracted API that is both easier to interact with when writing
tests and also provides a consistent interface for performing actions
on boards that may not have the same serial console semantics.

Signed-off-by: Sam Friedman <sam@golioth.io>