ci: parameterize hil tests
Use callable workflow to parameterize the hil test so that it
can be built for and run on different boards. Use a matrix
of boards to call the workflow. This strategy is used instead of
using a matrix in an uncallable workflow because the workflow needs
to be in 2 jobs (one for build in the cloud and one for test on
self-hosted runners) and the matrix strategy works only on jobs
not on workflows.

Signed-off-by: Sam Friedman <sam@golioth.io>