hil: zephyr: increase shell serial rx buffer
Increase the size of the serial receive ringbuffer for the shell, to
prevent the buffer from filling up and causing characters to be
dropped. This can happen during HIL tests because if the DUT is
running a higher priority thread, then the host PC can write to the
serial port faster than the DUT empties the ringbuffer. This isn't a
problem for the samples because humans don't type that fast, so this
change is only made for the HIL test app.

Signed-off-by: Sam Friedman <sam@golioth.io>