src: remove golioth_statistics module
The golioth_statistics module is used to track memory allocations
and detect memory leaks. It is a brittle, manual tool that requires
every allocation and free to be marked with a call into the module.
This clutters the code, is prone to breaking (if somebody forgets
to add a tracking call or adds it improperly), and can add overhead.
There are better ways to detect memory leaks including:

- using sanitizers
- using memory tools like valgrind
- using static analysis
- wrapping malloc and free to implement custom tracking
- using a gdb plugin

Signed-off-by: Sam Friedman <sam@golioth.io>