examples: zephyr: lightdb/set: wait for reboot before beginning test
Sometimes, the board manages to connect to Golioth and start setting
values in ligthdb state before the reboot occurs, which can cause the
test script to start waiting for input prematurely. Waiting for an
explicit "Booting" message before starting the test ensures that the
board is in a known state and ready to be tested.

Signed-off-by: Sam Friedman <sam@golioth.io>