treewide: un-typedef `golioth_client_t`
The golioth_client_t typedef was a void * which has 2
negative consequences: the `const` keyword won't work
correctly, and we lose type-checking at compile time.
We want to hide the members of the struct, not the fact
that it is a struct. So we can use an opaque struct pointer
instead of a void * custom type.

Signed-off-by: Sam Friedman <sam@golioth.io>