treewide: un-typedef `golioth_tls_auth_type_t`
We were using a typedef but still expecting the user to
interact with the type in a way that requires them to know
the underlying type. This obscures information without actually
encapsulating anything, making the code harder to read.

Also, rename to `golioth_auth_type` (without `tls`) so its a
little less specific.

Signed-off-by: Sam Friedman <sam@golioth.io>