coap_client: un-typedef `golioth_psk_credentials_t`
We were using a typedef but still expecting the user to
interact with the type in a way that requires them to know
the underlying type. This obscures information without actually
encapsulating anything, making the code harder to read.

Rename to `golioth_psk_credential` (not plural) to reflect
that it only holds one credential.

Signed-off-by: Sam Friedman <sam@golioth.io>