coap_client: `golioth_tls_credentials_t`
Signed-ofing a typedef but still expecting the user to
interact with the type in a way that requires them to know
the underlying type. This obscures information without actually
encapsulating anything, making the code harder to read.

Rename to `golioth_tls_credential` (singular) to reflect that
the struct only holds a single credential.

Signed-off-by: Sam Friedman <sam@golioth.io>