settings: remove settings struct from coap_client
The coap_client module should not have any knowledge of the
services built on top of it. Remove the settings struct from
the coap_client and instead create a new settings init function
that creates the settings struct, adds a pointer to the client,
and returns the struct to the user.

Signed-off-by: Sam Friedman <sam@golioth.io>