nrf91_lte_monitor: detect modem reset loop restriction
The nrf91 modem has a built in reset loop restriction that
activates when the device is reset 7 times in a row within
30 seconds of connecting to the network. This restriction
prevents the modem from connecting to the network for 30
minutes. We hit this frequently during CI. This commit
adds a notification when the reset loop restriction is
active.

Signed-off-by: Sam Friedman <sam@golioth.io>