nrf91_lte_monitor: automatically override reset loop restriction
Adds an option (default disabled) to automatically factory reset the
cellular modem when the modem reports that the reset loop restriction
is active. This will remove the reset loop restriction. This option
is NOT recommended for production deployments, but is useful in CI
where many reboots in rapid succession are expected and the reset
loop restriction frequently activates and interfers with tests.

Signed-off-by: Sam Friedman <sam@golioth.io>