ci: use auto-generated boards for twister tests
Don't provide the `GOLIOTH_DEVICE_NAME` environment variable to
the twister tests. This will cause the golioth pytest plugin to
automatically generate a temporary device in the console to use
for the duration of the test, after which it will be deleted.
Using automatically generated devices reduces the amount of manual
work needed to add a device into the HIL runners (we no longer need
to manually create devices) and will make it easier to move to
a different project or backend environment in the future.

Signed-off-by: Sam Friedman <sam@golioth.io>