ci: upgrade clang-format version
clang-format had been returning a 0 exit code even when it found
files that didn't match the coding standard. This is a bug in the
(rather old) version of clang-format that we had been using (14)
that is fixed in newer versions. Upgrading to clang-format 18
ensures that the CI check will fail if coding style violations
are found.

Signed-off-by: Sam Friedman <sam@golioth.io>